Implemented hook for measuring GPU times
